[TASK] Commit the csv files
These csv files came from searching through forge.typo3.org. Redmine had
a convenient option to download the search results as csv, so I used
that as the basis to find all of the relevant issues that I wanted to
mirror.

The final issue list and uri list is in the combined folder.